Attempt to stablish commit message convention
Still not enough content to really justify version control, tried to add some list contents to list_list and rearranged some stuff needlessly.

Figure and figcaption make it easier to read and look nicer on page, still unsure as to the full functionaity of the tags. StackOverflow suggestion.

Add CSS and JS files to site, give site a purpose.