Add content and restructure HTML
Renovating. Focusing on getting comfortable with VSCode while I
continue to create enough changes to continue justifying psuedo-
regular commits.

It's tough to have sensical commits when the pages are composed of
non-sense.

Moving towards using this as a working repository for links and info
that I'd liketo hold onto.